+++ nexus/src/app/silo.rs
@@ -1,0 +1,150 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+//! Silos, Users, and SSH Keys.
+
+use crate::authz;
+use crate::context::OpContext;
+use crate::db;
+use crate::db::identity::Resource;
+use crate::db::lookup::LookupPath;
+use crate::db::model::Name;
+use crate::db::model::SshKey;
+use crate::external_api::params;
+use omicron_common::api::external::CreateResult;
+use omicron_common::api::external::DataPageParams;
+use omicron_common::api::external::DeleteResult;
+use omicron_common::api::external::ListResultVec;
+use omicron_common::api::external::LookupResult;
+use uuid::Uuid;
+
+impl super::Nexus {
+    // Silos
+
+    pub async fn silo_create(
+        &self,
+        opctx: &OpContext,
+        new_silo_params: params::SiloCreate,
+    ) -> CreateResult<db::model::Silo> {
+        let silo = db::model::Silo::new(new_silo_params);
+        self.db_datastore.silo_create(opctx, silo).await
+    }
+
+    pub async fn silos_list_by_name(
+        &self,
+        opctx: &OpContext,
+        pagparams: &DataPageParams<'_, Name>,
+    ) -> ListResultVec<db::model::Silo> {
+        self.db_datastore.silos_list_by_name(opctx, pagparams).await
+    }
+
+    pub async fn silos_list_by_id(
+        &self,
+        opctx: &OpContext,
+        pagparams: &DataPageParams<'_, Uuid>,
+    ) -> ListResultVec<db::model::Silo> {
+        self.db_datastore.silos_list_by_id(opctx, pagparams).await
+    }
+
+    pub async fn silo_fetch(
+        &self,
+        opctx: &OpContext,
+        name: &Name,
+    ) -> LookupResult<db::model::Silo> {
+        let (.., db_silo) = LookupPath::new(opctx, &self.db_datastore)
+            .silo_name(name)
+            .fetch()
+            .await?;
+        Ok(db_silo)
+    }
+
+    pub async fn silo_delete(
+        &self,
+        opctx: &OpContext,
+        name: &Name,
+    ) -> DeleteResult {
+        let (.., authz_silo, db_silo) =
+            LookupPath::new(opctx, &self.db_datastore)
+                .silo_name(name)
+                .fetch_for(authz::Action::Delete)
+                .await?;
+        self.db_datastore.silo_delete(opctx, &authz_silo, &db_silo).await
+    }
+
+    // Users
+
+    pub async fn silo_user_fetch(
+        &self,
+        opctx: &OpContext,
+        silo_user_id: Uuid,
+    ) -> LookupResult<db::model::SiloUser> {
+        let (.., db_silo_user) = LookupPath::new(opctx, &self.datastore())
+            .silo_user_id(silo_user_id)
+            .fetch()
+            .await?;
+        Ok(db_silo_user)
+    }
+
+    // SSH Keys
+
+    pub async fn ssh_key_create(
+        &self,
+        opctx: &OpContext,
+        silo_user_id: Uuid,
+        params: params::SshKeyCreate,
+    ) -> CreateResult<db::model::SshKey> {
+        let ssh_key = db::model::SshKey::new(silo_user_id, params);
+        let (.., authz_user) = LookupPath::new(opctx, &self.datastore())
+            .silo_user_id(silo_user_id)
+            .lookup_for(authz::Action::CreateChild)
+            .await?;
+        assert_eq!(authz_user.id(), silo_user_id);
+        self.db_datastore.ssh_key_create(opctx, &authz_user, ssh_key).await
+    }
+
+    pub async fn ssh_keys_list(
+        &self,
+        opctx: &OpContext,
+        silo_user_id: Uuid,
+        page_params: &DataPageParams<'_, Name>,
+    ) -> ListResultVec<SshKey> {
+        let (.., authz_user) = LookupPath::new(opctx, &self.datastore())
+            .silo_user_id(silo_user_id)
+            .lookup_for(authz::Action::ListChildren)
+            .await?;
+        assert_eq!(authz_user.id(), silo_user_id);
+        self.db_datastore.ssh_keys_list(opctx, &authz_user, page_params).await
+    }
+
+    pub async fn ssh_key_fetch(
+        &self,
+        opctx: &OpContext,
+        silo_user_id: Uuid,
+        ssh_key_name: &Name,
+    ) -> LookupResult<SshKey> {
+        let (.., ssh_key) = LookupPath::new(opctx, &self.datastore())
+            .silo_user_id(silo_user_id)
+            .ssh_key_name(ssh_key_name)
+            .fetch()
+            .await?;
+        assert_eq!(ssh_key.name(), ssh_key_name);
+        Ok(ssh_key)
+    }
+
+    pub async fn ssh_key_delete(
+        &self,
+        opctx: &OpContext,
+        silo_user_id: Uuid,
+        ssh_key_name: &Name,
+    ) -> DeleteResult {
+        let (.., authz_user, authz_ssh_key) =
+            LookupPath::new(opctx, &self.datastore())
+                .silo_user_id(silo_user_id)
+                .ssh_key_name(ssh_key_name)
+                .lookup_for(authz::Action::Delete)
+                .await?;
+        assert_eq!(authz_user.id(), silo_user_id);
+        self.db_datastore.ssh_key_delete(opctx, &authz_ssh_key).await
+    }
+}
